Security module was divided into submodules
