+++ security/pom.xml
@@ -3,6 +3,10 @@
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
+    <modules>
+        <module>security-web</module>
+        <module>security-api</module>
+    </modules>
     <parent>
         <artifactId>text-adventure-base</artifactId>
         <groupId>org.text-adventure</groupId>
@@ -11,7 +15,7 @@
 
     <artifactId>security</artifactId>
     <version>1.0-SNAPSHOT</version>
-    <packaging>jar</packaging>
+    <packaging>pom</packaging>
 
     <build>
         <plugins>
