+++ security/security-api/src/main/java/org/my/adventure/security/api/SecurityConstants.java
@@ -1,0 +1,8 @@
+package org.my.adventure.security.api;
+
+/**
+ * Created by al on 25.02.2016.
+ */
+public class SecurityConstants {
+    public final static String ADMIN_ROLE = "admin";
+}
