+++ security/wildfly_configurations/standalone-full.xml
