+++ text-adventure-app/text-adventure-app-web/pom.xml
@@ -35,7 +35,7 @@
         </dependency>
         <dependency>
             <groupId>org.text-adventure</groupId>
-            <artifactId>security</artifactId>
+            <artifactId>security-web</artifactId>
             <version>1.0-SNAPSHOT</version>
             <type>jar</type>
         </dependency>
