+++ text-adventure-app/text-adventure-app-web/src/main/webapp/authorization/auth_form.xhtml
@@ -24,8 +24,7 @@
                 </p:inputText>
                 <p:message for="j_username" display="icon"/>
                 <p:outputLabel for="j_password" value="Пароль" />
-                <p:password id="j_password" required="true"
-                            tabindex="2"
+                <p:password id="j_password" required="true" tabindex="2"
                             validatorMessage="Длина пароля от 6 до 50 символов"
                             requiredMessage="Длина пароля от 6 до 50 символов">
                     <f:validateLength minimum="6" maximum="50" />
