+++ text-adventure-app/text-adventure-app-web/src/main/webapp/authorization/login_error.xhtml
@@ -10,7 +10,7 @@
     <!--<ui:define name="menu"/>-->
     <ui:define name="content">
         <div align="center" style="margin-top: 10%">
-            <h3 align="center">Неверное имя пользователя или пароль, попробуйте войти ещё раз.</h3>
+            <h3 align="center" style="color: red">Неверное имя пользователя или пароль, попробуйте войти ещё раз.</h3>
             <ui:include src="/authorization/auth_form.xhtml"/>
         </div>
     </ui:define>
